Rating and project completion

+++ AllocatrApi/Models/Review.cs
@@ -13,8 +13,10 @@ public class Review
     public Guid AllocatProfileId { get; set; }
     public AllocatProfile AllocatProfile { get; set; } = null!;
 
-    public decimal Rating { get; set; }
+    public int Rating { get; set; }
+
     public string? Comment { get; set; }
+
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set; }
 }
